+++ src/cmd/cmd_sym.mli
@@ -2,6 +2,12 @@
 (* Copyright © 2021-2024 OCamlPro *)
 (* Written by the Owi programmers *)
 
+type fail_mode =
+  [ `Trap_only
+  | `Assertion_only
+  | `Both
+  ]
+
 val cmd :
      bool
   -> bool
@@ -11,6 +17,7 @@ val cmd :
   -> bool
   -> bool
   -> bool
+  -> fail_mode
   -> Fpath.t
   -> Smtml.Solver_dispatcher.solver_type
   -> Fpath.t list